EMS-34 Init branch
[+] Add Device Connection Manager
[+] Add Table Reader from SSH result
[+] Add Interface and Port parser

+++ src/main/java/vn/com/tma/emsbackend/dto/NetworkDeviceDto.java
@@ -22,7 +22,7 @@ public class NetworkDeviceDto {
 
     private String model;
 
-    private long port;
+    private int port;
 
     private long credentialId;
